Add schedule page complete
